Fix: minor modularisation and code clean up
