+++ go-wasm-bridge/host_do_api_call.go
@@ -75,13 +75,7 @@ func (h *DoApiCall) callback(
 	}
 
 	responseStr := string(body)
-	wasmInput := WasmInput{
-		Caller:        caller,
-		AllocFunction: h.allocFunc,
-		Memory:        memory,
-		OutputValue:   responseStr,
-	}
-	err = UpdateDataToWASM(&wasmInput, outputArgs)
+	err = UpdateDataToWASM(caller, h.allocFunc, responseStr, outputArgs)
 	if err != nil {
 		fmt.Println("Failed to update data to WASM", err)
 		return utils.HandleError(err.Error())
