+++ go-wasm-bridge/host_do_transfer_ft_api_call.go
@@ -138,13 +138,7 @@ func (h *DoTransferFTApiCall) callback(
 	}
 
 	responseStr := "success"
-	wasmInput := WasmInput{
-		Caller:        caller,
-		AllocFunction: h.allocFunc,
-		Memory:        memory,
-		OutputValue:   responseStr,
-	}
-	err = UpdateDataToWASM(&wasmInput, outputArgs)
+	err = UpdateDataToWASM(caller, h.allocFunc, responseStr, outputArgs)
 	if err != nil {
 		fmt.Println("Failed to update data to WASM", err)
 		return utils.HandleError(err.Error())
