+++ go-wasm-bridge/wasm.go
@@ -10,7 +10,6 @@ import (
 	"os"
 
 	"github.com/bytecodealliance/wasmtime-go"
-	"github.com/rubixchain/rubix-wasm/go-wasm-bridge/utils"
 )
 
 // WasmModule encapsulates the WASM module and its associated functions.
@@ -30,12 +29,12 @@ type WasmModule struct {
 
 // WasmInput struct is defined beacuse, or else we need to provide 5 input params
 // Deined for better readability
-type WasmInput struct {
-	Caller        *wasmtime.Caller
-	AllocFunction *wasmtime.Func
-	Memory        *wasmtime.Memory
-	OutputValue   string
-}
+// type WasmInput struct {
+// 	Caller        *wasmtime.Caller
+// 	AllocFunction *wasmtime.Func
+// 	Memory        *wasmtime.Memory
+// 	OutputValue   string
+// }
 
 // caller *wasmtime.Caller, allocFunction *wasmtime.Func, memory *wasmtime.Memory, outputValue string, outputArg *utils.WasmArgInfo
 // WasmModuleOption allows us to configure WasmModule
@@ -264,79 +263,4 @@ func (w *WasmModule) CallFunction(args string) (string, error) {
 	}
 
 	return contractOutputStr, nil
-}
-
-func ExtractDataFromWASM(caller *wasmtime.Caller, inputArg *utils.WasmArgInfo) ([]byte, *wasmtime.Memory, error) {
-	// Access memory from the caller
-	memory := caller.GetExport("memory").Memory()
-	if memory == nil {
-		errMsg := "memory export not found"
-		return nil, nil, errors.New(errMsg)
-	}
-	//	h.memory = memory // Assign memory to Host struct for future use
-
-	// Read the input string from WASM memory
-	wasmMemory := memory.UnsafeData(caller)
-	if wasmMemory == nil {
-		errMsg := "Failed to get memory data"
-		return nil, nil, errors.New(errMsg)
-	}
-
-	// Convert pointers to int for slicing
-	inputStart := int(inputArg.DataPtr)
-	inputEnd := inputStart + int(inputArg.DataPtrSize)
-
-	// Validate memory bounds
-	if inputStart < 0 || inputEnd > len(wasmMemory) {
-		errMsg := "input exceeds memory bounds"
-		return nil, nil, errors.New(errMsg)
-	}
-
-	// Extract input bytes and convert to string
-	inputBytes := wasmMemory[inputStart:inputEnd]
-
-	return inputBytes, memory, nil
-}
-
-func UpdateDataToWASM(wasmInput *WasmInput, outputArg *utils.WasmArgInfo) error {
-	outputValueLen := int32(len(wasmInput.OutputValue))
-
-	// Allocating memory for output
-	result, err := wasmInput.AllocFunction.Call(wasmInput.Caller, outputValueLen)
-	if err != nil {
-		return err
-	}
-	wasmMemory := wasmInput.Memory.UnsafeData(wasmInput.Caller)
-	if wasmMemory == nil {
-		errMsg := "Failed to get memory data"
-		return fmt.Errorf("%s", errMsg)
-	}
-
-	// Type Cast the allocated pointer
-	respPtr, ok := result.(int32)
-	if !ok {
-		errMsg := "Alloc function did not return i32"
-		return fmt.Errorf("%s", errMsg)
-	}
-
-	// Get memory size to ensure we don't write out of bounds
-	memSize := wasmInput.Memory.DataSize(wasmInput.Caller)
-	if uint32(respPtr)+uint32(outputValueLen) > uint32(memSize) {
-		errMsg := "Response exceeds memory bounds"
-		return fmt.Errorf("%s", errMsg)
-	}
-
-	// Write response bytes to allocated memory
-	copy(wasmMemory[respPtr:], []byte(wasmInput.OutputValue))
-
-	respPtrPtr := outputArg.DataPtr
-	respLenPtr := outputArg.DataPtrSize
-
-	// Write the response pointer back to WASM memory using Little Endian encoding
-	binary.LittleEndian.PutUint32(wasmMemory[respPtrPtr:], uint32(respPtr))
-
-	// Write the response length back to WASM memory using Little Endian encoding
-	binary.LittleEndian.PutUint32(wasmMemory[respLenPtr:], uint32(outputValueLen))
-
-	return nil
 }
